add time slider
The play function works, but when stepping in between there are problems with the opacity.
The "step" (or sliding the slider with the mouse)
can hit an intermediate stage where two colors
are active at the same time and since we are
using opacity, both contribute to the visible image.