Rochester left the stretch code
The current DOER list does not list Rochster as a stretch code community any longer, though previous records did form 1/1/2019 on. The town website has select board meeting minutes  from 5/15/23 that put withdrawing from the stretch code on the Agenda for the 5/22/23 town meeting.
Unfortunately, I can't find the minutes of that meeting on the town website at the time of writing. However, there are tow newspaper articles, that also confirm that the town decided to leave the Stretch code on that date, so I'll use 5/22/23 as the date.

https://sippican.theweektoday.com/article/rochester-voters-leave-%E2%80%98green-communities%E2%80%99-program-first-state/63969 (note type ion date 2024->2023 in the article)
https://www.wanderer.com/news/one-vote-to-decide-green-communities-stretch-code/